Refactor HomePageViewModel to implement Logout functionality and remove duplicate Logout command from UserRoleEntryViewModel
